Fix use case that client waits for server greeting
This is an intrusive change. A conn->flags bit is occupied. This
was avoided intentionally before.

Formerly, the shadowsocks 'connect dest' command was only sent
after client sent some data. For http and SSL connections, that
was good enough. But SMTP had problem with the behaviour.

The code base is also rebased upon haproxy 1.5.16. haproxy 1.5.16
fixed a bug which causes busy loop and 100% cpu usage.